multi file changes 6
-added code of Grid Axes into its own .h and .cpp file
-setup MVP matrix calculation in display() and sent the mat4 for Rendering Grid Boxes.
-Need to do the same for all others to remove duplicate code.

TODO:
-final code cleanup and commented code needed
