+++ VolumeApp.cpp
@@ -25,6 +25,7 @@
 #include "./src/effects/Basic_TextureSlicing/BasicTextureSlicing.h"
 #include "./src/effects/RayCasting/RayCasting.h"
 #include "./src/effects/MarchingTetrahedra/MarchingTetrahedra.h"
+#include "./src/effects/GridBoxes/GridBoxes.h"
 
 
 
@@ -128,23 +129,6 @@ const wchar_t* effectNames[] =
 
 const wchar_t* textHolder = L"";
 
-GLuint shaderProgramObject_Grid;
-GLuint mvpUniform_GridObject;
-
-glm::vec3 vertexList[8] = { 
-						   // Minus Z-Vertices
-						   glm::vec3(-0.5,-0.5,-0.5), // Left Bottom // 0th
-						   glm::vec3(0.5,-0.5,-0.5),  // Right Bottom // 1st
-						   glm::vec3(0.5, 0.5,-0.5),  // Right Top
-						   glm::vec3(-0.5, 0.5,-0.5), // Left Top
-
-						   // Plus Z-Vertices
-						   glm::vec3(-0.5,-0.5, 0.5), // Left Bottom
-						   glm::vec3(0.5,-0.5, 0.5),  // Right Bottom
-						   glm::vec3(0.5, 0.5, 0.5),  // Right Top
-						   glm::vec3(-0.5, 0.5, 0.5)  // Left Top
-							};
-
 
 
 //* Object 1 Data Info
@@ -173,41 +157,7 @@ GLfloat xMouseValue_NDC = 0.0f;
 GLfloat yMouseValue_NDC = 0.0f;
 
 
-//!  VOLUME RENDERING Function Prototypes:
 
-
-// Grid and Axes Rendering Prototype:
-int LoadGridObject_Shader(int width, int depth);
-void setup_Axes_BufferObjects(void);
-void Update_Volume_Box_Axes(void);
-void Uninitialize_Grid(void);
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-GLuint VAO_Volume_Axes;
-GLuint VBO_Volume_Axes;
-
-void Update_Volume_Box_Axes(void);
-void Render_Volume_Box_Axes(glm::mat4);
 
 
 
@@ -1545,6 +1495,14 @@ void display(void)
 
 	glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
+	glm::mat4 ModelViewMatrix = glm::translate(glm::mat4(1.0f), glm::vec3(0.0f, 0.0f, dist));
+
+	ModelViewMatrix = glm::rotate(ModelViewMatrix, glm::radians(rotationX), glm::vec3(1.0f, 0.0f, 0.0f));
+	ModelViewMatrix = glm::rotate(ModelViewMatrix, glm::radians(rotationY), glm::vec3(0.0f, 1.0f, 0.0f));
+	ModelViewMatrix = glm::rotate(ModelViewMatrix, rotationZ, glm::vec3(0.0f, 0.0f, 1.0f));
+
+	glm::mat4 modelViewProjectionMatrix = perspectiveProjMatrix_glm * ModelViewMatrix;
+
 
 
 	switch (iEffectUsed)
@@ -1552,6 +1510,7 @@ void display(void)
 		case 0:
 			glClearColor(0.75f, 0.75f, 0.75f, 0.0f);
 			Render_Basic_Volume();
+			Render_Volume_Box_Axes(modelViewProjectionMatrix);
 			Update_Basic_Volume();
 			if (bMouseClicked || bSliceUpdate)
 			{
@@ -1565,6 +1524,7 @@ void display(void)
 			glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
 			Render_Raycasting_Output();
+			Render_Volume_Box_Axes(modelViewProjectionMatrix);
 			Update_Raycasting_Output();
 			if (bMouseClicked || bSliceUpdate)
 			{
@@ -1578,6 +1538,7 @@ void display(void)
 			glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
 			Render_IsoSurface_Output();
+			Render_Volume_Box_Axes(modelViewProjectionMatrix);
 			Update_IsoSurface_Output();
 			if (bMouseClicked || bSliceUpdate)
 			{
@@ -1591,6 +1552,7 @@ void display(void)
 			glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
 			Render_ColormapClassification_Output();
+			Render_Volume_Box_Axes(modelViewProjectionMatrix);
 			Update_ColormapClassification_Output();
 			if (bMouseClicked || bSliceUpdate)
 			{
@@ -1604,6 +1566,7 @@ void display(void)
 			glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 
 			Render_MarchingTetrahedra();
+			Render_Volume_Box_Axes(modelViewProjectionMatrix);
 			Update_MarchingTetrahedra();
 			if (bMouseClicked || bSliceUpdate)
 			{
@@ -1719,127 +1682,6 @@ void uninitialize(void)
 }
 
 
-void Update_Volume_Box_Axes(void)
-{
-	// code:
-
-	glm::vec3 vertexList[24] = {
-
-		// FRONT FACE: All Plus Z-Vertices
-		glm::vec3(fXMinus_SideFace,fYMinus_BottomFace, fZPlus_FrontFace),				// 5. Left Bottom
-		glm::vec3(fXPlus_SideFace,fYMinus_BottomFace, fZPlus_FrontFace),	// 6. Right Bottom
-		glm::vec3(fXPlus_SideFace, fYPlus_TopFace, fZPlus_FrontFace),		// 7. Right Top
-		glm::vec3(fXMinus_SideFace, fYPlus_TopFace, fZPlus_FrontFace),					// 8. Left Top
-
-		// RIGHT FACE : All Plus X-Vertices:
-		glm::vec3(fXPlus_SideFace, fYPlus_TopFace, fZPlus_FrontFace),
-		glm::vec3(fXPlus_SideFace, fYPlus_TopFace,fZMinus_BackFace),
-		glm::vec3(fXPlus_SideFace,fYMinus_BottomFace,fZMinus_BackFace),
-		glm::vec3(fXPlus_SideFace,fYMinus_BottomFace, fZPlus_FrontFace),
-
-		// TOP FACE: All Plus Y-Vertices
-		glm::vec3(fXMinus_SideFace, fYPlus_TopFace,fZMinus_BackFace),
-		glm::vec3(fXPlus_SideFace, fYPlus_TopFace,fZMinus_BackFace),
-		glm::vec3(fXPlus_SideFace, fYPlus_TopFace, fZPlus_FrontFace),
-		glm::vec3(fXMinus_SideFace, fYPlus_TopFace, fZPlus_FrontFace),
-
-		// BOTTOM FACE: All Minus Y-Vertices
-		glm::vec3(fXMinus_SideFace, fYMinus_BottomFace,fZMinus_BackFace),
-		glm::vec3(fXPlus_SideFace, fYMinus_BottomFace,fZMinus_BackFace),
-		glm::vec3(fXPlus_SideFace, fYMinus_BottomFace, fZPlus_FrontFace),
-		glm::vec3(fXMinus_SideFace, fYMinus_BottomFace, fZPlus_FrontFace),
-
-		// LEFT FACE : All Minus X-Vertices:
-		glm::vec3(fXMinus_SideFace, fYPlus_TopFace, fZPlus_FrontFace),
-		glm::vec3(fXMinus_SideFace, fYPlus_TopFace,fZMinus_BackFace),
-		glm::vec3(fXMinus_SideFace,fYMinus_BottomFace,fZMinus_BackFace),
-		glm::vec3(fXMinus_SideFace,fYMinus_BottomFace, fZPlus_FrontFace),
-
-		// BACK FACE: All Minus Z-Vertices
-		glm::vec3(fXMinus_SideFace,fYMinus_BottomFace, fZMinus_BackFace),				// 5. Left Bottom
-		glm::vec3(fXPlus_SideFace,fYMinus_BottomFace, fZMinus_BackFace),	// 6. Right Bottom
-		glm::vec3(fXPlus_SideFace, fYPlus_TopFace, fZMinus_BackFace),		// 7. Right Top
-		glm::vec3(fXMinus_SideFace, fYPlus_TopFace, fZMinus_BackFace),					// 8. Left Top
-
-	};
-
-
-	glBindBuffer(GL_ARRAY_BUFFER, VBO_Volume_Axes);
-	glBufferData(GL_ARRAY_BUFFER, sizeof(vertexList), &(vertexList[0].x), GL_DYNAMIC_DRAW);
-	glBindBuffer(GL_ARRAY_BUFFER, 0);
-
-}
-
-void Render_Volume_Box_Axes(glm::mat4 MVPMatrix_)
-{
-	// code:
-
-	glUseProgram(shaderProgramObject_Grid);
-	{
-		// Front Face
-		glUniformMatrix4fv(mvpUniform_GridObject, 1, GL_FALSE, glm::value_ptr(MVPMatrix_));
-		glUniform4f(glGetUniformLocation(shaderProgramObject_Grid, "u_Color"), 1.0f, 0.0f, 0.0f, 1.0f);
-		glBindVertexArray(VAO_Volume_Axes);
-		glDrawArrays(GL_LINE_LOOP, 0, 4);
-		glBindVertexArray(0);
-
-		// Right Face
-		glUniformMatrix4fv(mvpUniform_GridObject, 1, GL_FALSE, glm::value_ptr(MVPMatrix_));
-		glUniform4f(glGetUniformLocation(shaderProgramObject_Grid, "u_Color"), 0.0f, 1.0f, 0.0f, 1.0f);
-		glBindVertexArray(VAO_Volume_Axes);
-		glDrawArrays(GL_LINE_LOOP, 4, 4);
-		glBindVertexArray(0);
-
-		// Top Face
-		glUniformMatrix4fv(mvpUniform_GridObject, 1, GL_FALSE, glm::value_ptr(MVPMatrix_));
-		glUniform4f(glGetUniformLocation(shaderProgramObject_Grid, "u_Color"), 0.0f, 0.0f, 1.0f, 1.0f);
-		glBindVertexArray(VAO_Volume_Axes);
-		glDrawArrays(GL_LINE_LOOP, 8, 4);
-		glBindVertexArray(0);
-
-		// Bottom Face
-		glUniformMatrix4fv(mvpUniform_GridObject, 1, GL_FALSE, glm::value_ptr(MVPMatrix_));
-		glUniform4f(glGetUniformLocation(shaderProgramObject_Grid, "u_Color"), 1.0f, 1.0f, 0.0f, 1.0f);
-		glBindVertexArray(VAO_Volume_Axes);
-		glDrawArrays(GL_LINE_LOOP, 12, 4);
-		glBindVertexArray(0);
-
-		// Left Face
-		glUniformMatrix4fv(mvpUniform_GridObject, 1, GL_FALSE, glm::value_ptr(MVPMatrix_));
-		glUniform4f(glGetUniformLocation(shaderProgramObject_Grid, "u_Color"), 0.0f, 1.0f, 1.0f, 1.0f);
-		glBindVertexArray(VAO_Volume_Axes);
-		glDrawArrays(GL_LINE_LOOP, 16, 4);
-		glBindVertexArray(0);
-
-		// Back Face
-		glUniformMatrix4fv(mvpUniform_GridObject, 1, GL_FALSE, glm::value_ptr(MVPMatrix_));
-		glUniform4f(glGetUniformLocation(shaderProgramObject_Grid, "u_Color"), 1.0f, 0.0f, 1.0f, 1.0f);
-		glBindVertexArray(VAO_Volume_Axes);
-		glDrawArrays(GL_LINE_LOOP, 20, 4);
-		glBindVertexArray(0);
-	}
-	glUseProgram(0);
-
-
-}
-void Uninitialize_Grid(void)
-{
-	// code:
-	if (VBO_Volume_Axes)
-	{
-		glDeleteBuffers(1, &VBO_Volume_Axes);
-		VBO_Volume_Axes  = 0;
-	}
-	if (VAO_Volume_Axes)
-	{
-		glDeleteVertexArrays(1, &VAO_Volume_Axes);
-		VAO_Volume_Axes = 0;
-	}
-
-	Uninitialize_ShaderProgramObject(shaderProgramObject_Grid);
-}
-
-
 
 
 
@@ -1940,209 +1782,6 @@ int LoadVolumeData(void)
 	return 0;
 }
 
-int LoadGridObject_Shader(int width, int depth)
-{
-	// prototype:
-	void uninitialize(void);
-
-	// local:
-
-	GLuint vertexShaderObject;
-	GLuint fragmentShaderObject;
-
-	GLint status;
-	GLint infoLogLength;
-	char *Log = NULL;
-
-	// code:
-
-	///* //////////////////// # VERTEX SHADER # ////////////////////////
-	const GLchar *vertexShaderSource = R"(
-			
-			#version 460 core
-			
-			layout (location = 0) in vec3 aPosition;
-			uniform mat4 u_MVPMatrix;
-			void main()
-			{
-				gl_Position = u_MVPMatrix * vec4(aPosition.xyz, 1.0);
-			}
-	
-			)";
-
-	vertexShaderObject = glCreateShader(GL_VERTEX_SHADER);
-	glShaderSource(vertexShaderObject, 1, (const GLchar **)&vertexShaderSource, NULL);
-
-	glCompileShader(vertexShaderObject);
-	glGetShaderiv(vertexShaderObject, GL_COMPILE_STATUS, &status);
-
-	if (status == GL_FALSE)
-	{
-		glGetShaderiv(vertexShaderObject, GL_INFO_LOG_LENGTH, &infoLogLength);
-		if (infoLogLength > 0)
-		{
-			Log = (char *)malloc(infoLogLength);
-			if (Log != NULL)
-			{
-				GLsizei written;
-				glGetShaderInfoLog(vertexShaderObject, infoLogLength, &written, Log);
-				fprintf(gpFile, " GridObject Vertex Shader Compilation Log : %s\n", Log);
-
-				free(Log);
-				Log = NULL;
-				uninitialize();
-				return -1;
-			}
-		}
-	}
-	else
-	{
-		fprintf(gpFile, "GridObject Success at  Vertex Shader Compilation \n");
-	}
-
-
-
-	//* ///////////////////// # FRAGMENT SHADER # ////////////////////////
-	const GLchar *fragmentShaderSource = R"(
-			
-			#version 460 core
-
-			uniform vec4 u_Color;
-			out vec4 FragColor;		
-			void main(void)
-			{             
-				FragColor = u_Color ;
-			}
-
-			)";
-	fragmentShaderObject = glCreateShader(GL_FRAGMENT_SHADER);
-	glShaderSource(fragmentShaderObject, 1, (const GLchar **)&fragmentShaderSource, NULL);
-
-	glCompileShader(fragmentShaderObject);
-	glGetShaderiv(fragmentShaderObject, GL_COMPILE_STATUS, &status);
-
-	if (status == GL_FALSE)
-	{
-		glGetShaderiv(fragmentShaderObject, GL_INFO_LOG_LENGTH, &infoLogLength);
-		if (infoLogLength > 0)
-		{
-			Log = (char *)malloc(infoLogLength);
-			if (Log != NULL)
-			{
-				GLsizei written;
-				glGetShaderInfoLog(fragmentShaderObject, infoLogLength, &written, Log);
-				fprintf(gpFile, "GridObject  FRAGMENT Shader Compilation Log : %s\n", Log);
-				free(Log);
-				Log = NULL;
-				uninitialize();
-				return -1;
-			}
-		}
-	}
-	else
-	{
-		fprintf(gpFile, "GridObject Success at  FRAGMENT Shader Compilation \n");
-	}
-
-	shaderProgramObject_Grid = glCreateProgram();
-
-
-	glAttachShader(shaderProgramObject_Grid, vertexShaderObject);
-	glAttachShader(shaderProgramObject_Grid, fragmentShaderObject);
-
-
-	glBindAttribLocation(shaderProgramObject_Grid, AMC_ATTRIBUTE_POSITION, "aPosition");
-
-	infoLogLength=0;
-	Log = NULL;
-	glLinkProgram(shaderProgramObject_Grid);
-	glGetProgramiv(shaderProgramObject_Grid, GL_LINK_STATUS, &status);
-	if (status == GL_FALSE)
-	{
-		glGetProgramiv(shaderProgramObject_Grid, GL_INFO_LOG_LENGTH, &infoLogLength);
-		if (infoLogLength > 0)
-		{
-			Log = (char*)malloc(infoLogLength);
-			if (Log != NULL)
-			{
-				GLsizei written;
-				glGetProgramInfoLog(shaderProgramObject_Grid, infoLogLength, &written, Log);
-				fprintf(gpFile, "GridObject  Shader Compilation Log : %s\n", Log);
-				free(Log);
-				uninitialize();
-				return -1;
-			}
-		}
-	}
-	else
-	{
-		fprintf(gpFile, "GridObject  Linking Successful \n");
-	}
-
-	mvpUniform_GridObject = glGetUniformLocation(shaderProgramObject_Grid, "u_MVPMatrix");
-
-	
-	return 0;
-}
-
-
-
-void setup_Axes_BufferObjects(void)
-{
-	// code:
-
-
-	glGenVertexArrays(1, &VAO_Volume_Axes);
-	glBindVertexArray(VAO_Volume_Axes);
-	{
-		glm::vec3 vertexList[24] = {
-			// FRONT FACE: All Plus Z-Vertices
-			glm::vec3(-0.5f,-0.5f, fZPlus_FrontFace), // 5. Left Bottom
-			glm::vec3(fXPlus_SideFace,-0.5f, fZPlus_FrontFace),  // 6. Right Bottom
-			glm::vec3(fXPlus_SideFace, fYPlus_TopFace, fZPlus_FrontFace),  // 7. Right Top
-			glm::vec3(-0.5f, fYPlus_TopFace, fZPlus_FrontFace),  // 8. Left Top
-
-			// RIGHT SIDE FACE : All Plus X-Vertices:
-			glm::vec3(fXPlus_SideFace, 0.5f, 0.5f),
-			glm::vec3(fXPlus_SideFace, 0.5f,-0.5f),
-			glm::vec3(fXPlus_SideFace,-0.5f,-0.5f),
-			glm::vec3(fXPlus_SideFace,-0.5f, 0.5f),
-
-			// TOP FACE : All Plus Y-Vertices:
-			glm::vec3(-0.5f, fYPlus_TopFace,-0.5f),
-			glm::vec3(0.5f, fYPlus_TopFace,-0.5f),
-			glm::vec3(0.5f, fYPlus_TopFace, 0.5f),
-			glm::vec3(-0.5f, fYPlus_TopFace, 0.5f),
-
-			// BOTTOM FACE : All Minus  Y-Vertices:
-			glm::vec3(-0.5f, fYMinus_BottomFace,-0.5f),
-			glm::vec3(0.5f, fYMinus_BottomFace,-0.5f),
-			glm::vec3(0.5f, fYMinus_BottomFace, 0.5f),
-			glm::vec3(-0.5f, fYMinus_BottomFace, 0.5f),
-
-			// BACK FACE: All Minus Z-Vertices
-			glm::vec3(-0.5f,-0.5f, fZMinus_BackFace),
-			glm::vec3(0.5f,-0.5f, fZMinus_BackFace),
-			glm::vec3(0.5f,  0.5f, fZMinus_BackFace),
-			glm::vec3(-0.5f,  0.5f, fZMinus_BackFace),
-
-			// LEFT SIDE FACE: All Minus X-Vertices
-			glm::vec3(fXMinus_SideFace, 0.5f, 0.5f),
-			glm::vec3(fXMinus_SideFace, 0.5f,-0.5f),
-			glm::vec3(fXMinus_SideFace,-0.5f,-0.5f),
-			glm::vec3(fXMinus_SideFace,-0.5f, 0.5f),
-
-		};
-
-		glGenBuffers(1, &VBO_Volume_Axes);
-		glBindBuffer(GL_ARRAY_BUFFER, VBO_Volume_Axes);
-		glBufferData(GL_ARRAY_BUFFER, sizeof(vertexList), NULL, GL_DYNAMIC_DRAW);
-		glEnableVertexAttribArray(AMC_ATTRIBUTE_POSITION);
-		glVertexAttribPointer(AMC_ATTRIBUTE_POSITION, 3, GL_FLOAT, GL_FALSE, 0, 0);
-		glBindBuffer(GL_ARRAY_BUFFER, 0);
-	}
-	glBindVertexArray(0);
-}
 
 
 bool LoadVolume_MT(void)
